refactor: ExchangeTokenStoring own file/extension; lowercase UITest ids + MARKs; reactive name prompt; trim/preview polish
