+++ Shared/Exchange/ExchangeTokenStore.swift
@@ -1,21 +1,11 @@
 import Foundation
 
-/// Per-account read-only-token storage seam. The production conformer is
-/// `ExchangeTokenStore` (keychain); test code injects a save-throwing
-/// double so `ExchangeAccountCreationLogic`'s rollback path can be
-/// exercised without keychain entitlements.
-protocol ExchangeTokenStoring: Sendable {
-  func save(token: String, for accountId: UUID) throws
-  func token(for accountId: UUID) throws -> String?
-  func delete(for accountId: UUID)
-}
-
 /// Per-account keychain storage for an exchange account's read-only access
 /// token. Each Moolah account gets its own keychain row keyed by account id,
 /// in the same env-scoped `apiKeys` service the Alchemy/CoinGecko keys use.
 /// Production uses the iCloud-synced keychain so the token follows the user
 /// across devices (the token is a secret and must never enter the DB/CloudKit).
-struct ExchangeTokenStore: ExchangeTokenStoring, Sendable {
+struct ExchangeTokenStore: Sendable {
   private let synchronizable: Bool
 
   /// - Parameter synchronizable: When `true` (the production default), the
@@ -32,7 +22,9 @@ struct ExchangeTokenStore: ExchangeTokenStoring, Sendable {
       account: "exchange-token-\(accountId.uuidString)",
       synchronizable: synchronizable)
   }
+}
 
+extension ExchangeTokenStore: ExchangeTokenStoring {
   func save(token: String, for accountId: UUID) throws {
     try keychainStore(for: accountId).saveString(token)
   }
